requirements.txt
discord.py[voice]==2.3.2
python-dotenv==1.0.1
requests==2.31.0
pynacl==1.5.0
google-generativeai==0.4.0
voicevox-client==0.2.0
ffmpeg-python==0.2.0
aiohttp==3.9.3
asyncio==3.4.3
SpeechRecognition==3.10.0
pyaudio==0.2.13
psutil==5.9.0
SQLAlchemy==2.0.0
pytest==7.4.0
pytest-asyncio==0.21.0